feat: Implement full sidewalk generation workflow
This commit implements the full sidewalk generation workflow as described in `data_assets/qgis_algorithm_description.md`.

The main changes are:
- The main algorithm logic has been moved from `main.py` to a new `full_sidewalkreator_algorithm.py` file to improve modularity.
- The `handle_sidewalk_tags` function now supports `sidewalk=left/right/yes/both` tags, implementing both exclusion and sure zones.
- A topological cleaning step has been added to the sidewalk splitting process to improve the quality of the generated geometries.
- The output generation has been updated to save additional auxiliary files (input polygon and intersection points).
- The hardcoded `curve_radius` in `draw_sidewalks_gdf` has been fixed.
- The tests have been updated to reflect these changes.
